eal: add counter size for efd clean
For virtual device, the rte_intr_handle struct is
initialized by the virtual device driver, including
the event fd assignment. If the event fd need to be
read for clean, an argument is required for the proper
event fd read.

This patch adds efd_counter_size in rte_intr_handle
struct to tell the rx interrupt process know the read
size.

Signed-off-by: Jingjing Wu <jingjing.wu@intel.com>