buildtools: zero elf info variable in pmdinfogen
This commit zeros out the elf_info struct at startup of the
pmdinfogen code. If it is not zeroed, later in the code gcc
produces "may be unused" prints. Clang does not report any
issue.

This commit enables a simplification in the meson build
system, removing the requirement for "-Wno-maybe-uninitialized".

Signed-off-by: Harry van Haaren <harry.van.haaren@intel.com>