net/i40e: fix compilation with -Og
The compilation with gcc-6.3.0 and EXTRA_CFLAGS=-Og gives the following
error:

  CC i40e_adminq.o
  i40e_adminq.c: In function ‘i40e_clean_arq_element’:
  i40e_adminq.c:1145:56: error: ‘ntu’ may be used uninitialized in
                         this function [-Werror=maybe-uninitialized]
     *pending = (ntc > ntu ? hw->aq.arq.count : 0) + (ntu - ntc);
                                                     ~~~~~^~~~~~

Depending on what is defined, ntu actually be used without being
initialized. Fix it by initializing it to 0, despite this is probably
not the proper fix. Moreover, the error is located in base/ directory,
which contains driver code common to several platforms (not only dpdk).

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>