eventdev: Add caps API and PMD callbacks for rte_event_eth_rx_adapter
The caps API allows application to get information
needed to configure the ethernet receive adapter for the eventdev and
ethdev pair.

The PMD callbacks are used by the rte_event_eth_rx_xxx() APIs to
configure and control the ethernet receive adapter if packet transfers
from the ethdev to eventdev is implemented in hardware.

For e.g., the ethdev, eventdev pairing maybe such that all of the
Eth Rx queues can only be connected to a single event queue, in
which case the application is required to pass in -1 as the queue id
when adding a receive queue to the adapter.

Signed-off-by: Nikhil Rao <nikhil.rao@intel.com>