net/fm10k: support switch restart on PF
For events indicating the change in the state of the switch manager. The
driver will restore the basic port configurations and then pass this event
to application so that the application can restore any additional
configurations if required. In this way, once the switch manager restarts,
the DPDK application can resume its network.

Signed-off-by: Xiao Wang <xiao.w.wang@intel.com>