net/mlx5: fix RSS hash update
Few bugs fixes in both configuration get and hash update where inputs are
not handled as expected by the ethdev layer.

RSS structure may not be totally usable, the PMD should try to take as most
information from it has it can when it is an hash update or it should try
to fill as most as possible in the configuration get.
This means that in the RSS configuration structure, the memory space for
the RSS hash key may not be present, but the PMD should consider the hash
field valid and process/set it.

Signed-off-by: Nelio Laranjeiro <nelio.laranjeiro@6wind.com>