lib/librte_flow_classify: remove table id parameter from apis
This patch removes table id parameter from all the flow
classify apis to reduce the complexity and and does some
cleanup of the code.

The validate api has been exposed as public api to allows
user to validate the flow before adding it to the
classifier.

The sample app and unit tests have been updated to accomodate
the apis changes.

Signed-off-by: Jasvinder Singh <jasvinder.singh@intel.com>